+++ src/main/java/ypjs/project/domain/Delivery.java
@@ -11,7 +11,7 @@ public class Delivery {
     @Id
     @GeneratedValue
     @Column(name = "delivery_id")
-    private int deliveryId;  //배송번호
+    private Long deliveryId;  //배송번호
 
     @OneToOne(mappedBy = "delivery", fetch = FetchType.LAZY)
     private Order orderId;  //주문번호
